fix: uf generated a router its own formatter rejected, and linted for legacy Flow
Two things stopped this repository from checking itself, and both were bugs in
uf rather than in the repository.

`uf fmt --check` failed on `docs/router.js` — a file uf writes. The generator
emitted `route < Path extends RoutePath > (`, with spaces the printer does not
put there, and omitted the trailing comma the printer adds. uf scaffolds a
project and then checks it with its own formatter, so this failed on code
nobody wrote, in every generated project. `the_generated_router_is_already_formatted`
runs the generator's output through `uf_fmt` and compares, which is a guarantee
rather than a promise to remember.

`uf lint` reported 152 `flow/ambiguous-object-type` errors against uf's own
packages for writing modern Flow. The rule is named after an ambiguity that no
longer exists: it dates from when a `.flowconfig` could set
`exact_by_default=false` and `{ a: b }` meant different things in different
projects. Flow has defaulted to exact since 2023 and now rejects that option as
deprecated, so `{ a: b }` is exact — and the `{| |}` the rule asks for is the
legacy spelling of what the plain braces already say. It is off by default now,
and still selectable for a codebase migrating from an older Flow with both
spellings in the tree.

The router generator was writing `{| |}` for exactly that rule, and says so in
a comment. It writes plain braces now, which is the same type in the spelling
uf tells everyone to use, and the exactness tests check what actually matters —
that there is no `...` — rather than which spelling was used.

`uf_router`'s tests move out of `lib.rs` into `tests.rs`, matching every other
crate in the workspace.

`uf fmt --check` now passes over this whole repository, so it is in `ci` and
has a job. `uf lint` is at 315 errors and is not: the largest groups are
`flow/unclear-type`, `flow/react-intrinsic-overlap` and `react/hooks-rules`,
and each needs looking at on its own terms — some are real findings, and some
may be wrong the way this one was.
